# openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "To-Do List API",
    "version": "1.0.0",
...
    "contact": {
      "name": "API Support",
      "email": "support@todolist-api.com",
      "url": "https://github.com/your-org/todolist-api"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.todolist.com/api",
      "description": "Production server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "required": ["id", "listId", "title", "completed", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the task (UUID v4)",
            "example": "b2c3d4e5-f6a7-8901-bcde-f23456789012"
          },
          "listId": {
            "type": "string",
            "format": "uuid",
            "description": "Foreign key reference to the parent list",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Title of the task (1-200 characters)",
            "example": "Complete API documentation"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "Optional detailed description of the task (max 1000 characters)",
            "example": "Write comprehensive Swagger documentation for all API endpoints",
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "description": "Optional deadline for the task (must be a future date when setting)",
            "example": "2025-07-10T17:00:00.000Z",
            "nullable": true
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the task has been completed",
            "example": false,
            "default": false
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was marked as completed (only set when completed = true)",
            "example": "2025-07-04T15:30:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was created",
            "example": "2025-07-04T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the task was last updated",
            "example": "2025-07-04T15:30:00.000Z"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/lists/{listId}/tasks": {
      "post": {
        "summary": "Create a new task in a list",
        "description": "Creates a new task within the specified list",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the list to add the task to"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Title of the task"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Optional description of the task"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional deadline for the task (future date)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get tasks for a specific list",
        "description": "Retrieves all tasks belonging to the specified list with optional filtering and sorting",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The list ID"
          },
          {
            "in": "query",
            "name": "includeCompleted",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Whether to include completed tasks"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": ["title", "createdAt", "updatedAt", "deadline"],
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "example": 5
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid UUID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{taskId}": {
      "put": {
        "summary": "Update a task",
        "description": "Updates an existing task with new data",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the task to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Updated title of the task"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Updated description of the task"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Updated deadline for the task"
                  },
                  "completed": {
                    "type": "boolean",
                    "description": "Updated completion status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Deletes an existing task",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the task to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}